+++ plugin-unity-agent/src/test/kotlin/jetbrains/buildServer/unity/detectors/UnityToolProviderTest.kt
@@ -2,12 +2,15 @@ package jetbrains.buildServer.unity.detectors
 
 import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.matchers.equals.shouldBeEqual
+import io.kotest.matchers.shouldBe
 import io.mockk.clearAllMocks
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.mockkStatic
 import jetbrains.buildServer.ExtensionHolder
 import jetbrains.buildServer.agent.*
+import jetbrains.buildServer.unity.DetectionMode
+import jetbrains.buildServer.unity.ProjectAssociatedUnityVersionIdentifier
 import jetbrains.buildServer.unity.UnityConstants
 import jetbrains.buildServer.unity.UnityEnvironment
 import jetbrains.buildServer.unity.UnityVersion
@@ -16,9 +19,8 @@ import jetbrains.buildServer.unity.util.unityVersionParam
 import jetbrains.buildServer.util.EventDispatcher
 import org.testng.annotations.BeforeMethod
 import org.testng.annotations.DataProvider
-import org.testng.annotations.Parameters
-import org.testng.annotations.Test
 import java.io.File
+import kotlin.test.Test
 import kotlin.test.assertEquals
 
 class UnityToolProviderTest {
@@ -30,6 +32,7 @@ class UnityToolProviderTest {
     private val toolsRegistry = mockk<ToolProvidersRegistry>()
     private val extensionHolder = mockk<ExtensionHolder>()
     private val eventDispatcher = mockk<EventDispatcher<AgentLifeCycleListener>>()
+    private val associatedVersionIdentifier = mockk<ProjectAssociatedUnityVersionIdentifier>()
 
     @BeforeMethod
     fun setUp() {
@@ -44,6 +47,7 @@ class UnityToolProviderTest {
         mockkStatic(runnerContext::unityVersionParam)
         every { runnerContext.unityRootParam() } returns null
         every { runnerContext.unityVersionParam() } returns null
+        every { associatedVersionIdentifier.identify(any()) } returns null
     }
 
     @Test
@@ -57,17 +61,6 @@ class UnityToolProviderTest {
     }
 
     @Test
-    fun `should throw exception when Unity detector was not created`() {
-        // arrange
-        every { unityDetectorFactory.unityDetector() } returns null
-        val provider = createInstance()
-
-        // act // assert
-        shouldThrowExactly<ToolCannotBeFoundException> { provider.getUnity("unity") }
-            .message?.shouldBeEqual("unity")
-    }
-
-    @Test
     fun `should detect Unity environment when Unity root param is provided`() {
         // arrange
         val provider = createInstance()
@@ -75,6 +68,7 @@ class UnityToolProviderTest {
         val unityVersion = UnityVersion(2023, 1, 1)
         every { runnerContext.unityRootParam() } returns unityRootParam
         every { unityDetector.getVersionFromInstall(File(unityRootParam)) } returns unityVersion
+        every { runnerContext.runnerParameters } returns emptyMap()
         every { unityDetector.getEditorPath(File(unityRootParam)) } returns File("$unityRootParam/Unity")
 
         // act
@@ -85,13 +79,79 @@ class UnityToolProviderTest {
     }
 
     @DataProvider
-    fun `should return latest Unity version if no one is specified explicitly params`(): Array<Array<Any>> = arrayOf(
+    fun `should return the latest (up to major or minor) Unity version when an exact version is not available params`(): Array<Array<Any>> = arrayOf(
+        arrayOf(
+            UnityVersion(2023, 2, 8),
+            listOf("2023.3.9", "2023.2.7", "2023.2.12"),
+            UnityVersion(2023, 2, 12),
+        ),
+        arrayOf(
+            UnityVersion(2023, null, null),
+            listOf("2023.1.2", "2023.3.7", "2023.3.9"),
+            UnityVersion(2023, 3, 9),
+        ),
+    )
+
+    @Test(dataProvider = "should return the latest (up to major or minor) Unity version when an exact version is not available params")
+    fun `should return the latest (up to major or minor) Unity version when an exact version is not available`(
+        specifiedVersion: UnityVersion,
+        versions: List<String>,
+        expectedVersion: UnityVersion,
+    ) {
+        // arrange
+        val provider = createInstance()
+        every { runnerContext.runnerParameters } returns mapOf(
+            UnityConstants.PARAM_DETECTION_MODE to DetectionMode.Auto.id,
+        )
+        every { runnerContext.unityVersionParam() } returns specifiedVersion
+        every { unityDetector.getEditorPath(any()) } returns mockk(relaxed = true)
+        every { agentConfiguration.configurationParameters } returns versions.associate {
+            "${UnityConstants.UNITY_CONFIG_NAME}${it}" to "/foo/${it}"
+        }
+        provider.agentStarted(mockk())
+
+        // act
+        val result = provider.getUnity("unity", runnerContext)
+
+        // assert
+        result.unityVersion shouldBe expectedVersion
+    }
+
+    @Test
+    fun `should pick up Unity version from the project settings if no one is specified explicitly`() {
+        // arrange
+        val provider = createInstance()
+        every { runnerContext.runnerParameters } returns mapOf(
+            UnityConstants.PARAM_DETECTION_MODE to DetectionMode.Auto.id,
+            UnityConstants.PARAM_PROJECT_PATH to "/bar"
+        )
+        every { runnerContext.unityVersionParam() } returns null
+        every { unityDetector.getEditorPath(any()) } returns mockk(relaxed = true)
+        every { runnerContext.workingDirectory } returns mockk(relaxed = true)
+        val expectedVersion = UnityVersion(2022, 3, 9)
+        every { agentConfiguration.configurationParameters } returns mapOf(
+            "${UnityConstants.UNITY_CONFIG_NAME}${expectedVersion}" to "/foo/${expectedVersion}",
+            "${UnityConstants.UNITY_CONFIG_NAME}2023.3.9" to "/foo/2023.3.9",
+            "${UnityConstants.UNITY_CONFIG_NAME}2021.3.9" to "/foo/2021.3.9",
+        )
+        every { associatedVersionIdentifier.identify(any()) } returns expectedVersion
+        provider.agentStarted(mockk())
+
+        // act
+        val result = provider.getUnity("unity", runnerContext)
+
+        // assert
+        result.unityVersion shouldBeEqual expectedVersion
+    }
+
+    @DataProvider
+    fun `should return latest Unity version if no one is specified explicitly and project settings unavailable params`(): Array<Array<Any>> = arrayOf(
         arrayOf(listOf("2020.3.38f1", "2020.3.43f1", "2020.3.32f1"), UnityVersion(2020, 3, 43)),
         arrayOf(listOf("2020.1.0f1", "2020.2.43f1", "2020.10.32f1"), UnityVersion(2020, 10, 32))
     )
 
-    @Test(dataProvider = "should return latest Unity version if no one is specified explicitly params")
-    fun `should return latest Unity version if no one is specified explicitly`(versions: List<String>, expectedVersion: UnityVersion) {
+    @Test(dataProvider = "should return latest Unity version if no one is specified explicitly and project settings unavailable params")
+    fun `should return latest Unity version if no one is specified explicitly and project settings unavailable`(versions: List<String>, expectedVersion: UnityVersion) {
         // arrange
         val provider = createInstance()
 
@@ -99,6 +159,12 @@ class UnityToolProviderTest {
             .associate { "${UnityConstants.UNITY_CONFIG_NAME}${it}" to "/Applications/Unity/Hub/Editor/${it}" }
         provider.agentStarted(mockk())
         every { runnerContext.unityRootParam() } returns null
+        every { runnerContext.runnerParameters } returns mapOf(
+            UnityConstants.PARAM_PROJECT_PATH to "/foo"
+        )
+        every { runnerContext.workingDirectory } returns mockk {
+            every { absolutePath } returns "/foo"
+        }
         every { unityDetector.getEditorPath(any()) } returns mockk(relaxed = true)
 
         // act
@@ -110,6 +176,7 @@ class UnityToolProviderTest {
 
     private fun createInstance() = UnityToolProvider(
         agentConfiguration,
+        associatedVersionIdentifier,
         unityDetectorFactory,
         toolsRegistry,
         extensionHolder,